+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Fixed
 
 - Fix vscode artifacts where arm64 package was not released.
+- IntelliJ plugin: if namespaces can't be accessed, use the default namespace
 
 ### Changed
 - Add `/home` to default file exclude list.